query부분 api로 분리
