+++ src/styles/styles.css
@@ -1,45 +1,3 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-.container {
-  width: 1400px;
-  margin: 20px auto;
-  columns: 4;
-  column-gap: 20px;
-}
-
-.container .box {
-  width: 100%;
-  margin-bottom: 10px;
-  break-inside: avoid;
-}
-
-.container .box img {
-  max-width: 100%;
-  border-radius: 15px;
-}
-
 .hidden-svg {
   display: none;
-}
-
-@media (max-width: 1200px) {
-  .container {
-    width: calc(100% - 40px);
-    columns: 3;
-  }
-}
-
-@media (max-width: 768px) {
-  .container {
-    columns: 2;
-  }
-}
-@media (max-width: 480px) {
-  .container {
-    columns: 1;
-  }
 }
